api/Model/model.py
from pydantic import BaseModel
from typing import Optional

class Token(BaseModel):
    access_token: str
    token_type: str

class TokenData(BaseModel):
    username: Optional[str] = None

class User(BaseModel):
    username: str
    password: str
    email: Optional[str] = None
    full_name: Optional[str] = None

class UserInLogin(BaseModel):
    username: str
    password: str

class Post(BaseModel):
    title: str
    description: Optional[str] = None