api/requirements.txt
fastapi
passlib[bcrypt]
python-jose[cryptography]
python-multipart
uvicorn
sqlmodel